+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.5
+
+- Enhancements styling
+- text parameter is removed
+- circular loader added
+
+## 0.0.4
+
+- SizedBox edit to customize size of the button
+
 ## 0.0.3
 
 - Example added